+++ src/wp-admin/includes/class-rollback-auto-update.php
@@ -348,12 +348,10 @@ class WP_Rollback_Auto_Update {
 		$ref_temp_backups->setValue( $rollback_updater, $temp_backup );
 
 		// Call Rollback's restore_temp_backup().
-		$restore_temp_backup = new ReflectionMethod( $rollback_updater, 'restore_temp_backup' );
-		$restore_temp_backup->invoke( $rollback_updater );
+		$rollback_updater->restore_temp_backup();
 
 		// Call Rollback's delete_temp_backup().
-		$delete_temp_backup = new ReflectionMethod( $rollback_updater, 'delete_temp_backup' );
-		$delete_temp_backup->invoke( $rollback_updater );
+		$rollback_updater->delete_temp_backup();
 
 		// TODO: remove before commit.
 		error_log( $this->handler_args['hook_extra']['plugin'] . ' rolled back' );