+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 [unreleased]
+* remove Reflection in `WP_Rollback_Auto_Update::cron_rollback()` as methods are public
 
 #### 5.3.2 / 2023-07-10
 * cleanup